WinUI - Start AboutDialog
